+++ uptick-dashboard/app/api/authOptions.js
@@ -1,0 +1,61 @@
+import CredentialsProvider from 'next-auth/providers/credentials';
+
+export const authOptions = {
+    providers: [
+        CredentialsProvider({
+            name: 'credentials',
+            credentials: {},
+
+            async authorize(credentials) {
+                const { username, password } = credentials;
+                const baseUrl = process.env.BASE_URL;
+
+                try {
+                    const response = await fetch(`${baseUrl}/auth/login`, {
+                        method: 'POST',
+                        headers: {
+                            'Content-Type': 'application/json'
+                        },
+                        body: JSON.stringify({
+                            username: username,
+                            password: password,
+                        }),
+                    });
+
+                    const user = await response.json();
+
+                    if (response.ok && user) {
+                        return user;
+                    } else {
+                        return null;
+                    }
+                } catch (error) {
+                    return null;
+                }
+            }
+        })
+    ],
+    session: {
+        strategy: 'jwt',
+    },
+    secret: process.env.NEXTAUTH_SECRET,
+    pages: {
+        signIn: '/',
+        signOut: '/',
+    },
+    callbacks: {
+        async jwt({ token, user }) {
+            if (user) {
+                token.accessToken = user.accessToken;
+                
+                // Decode token and get the payload
+                // const decodedToken = JSON.parse(Buffer.from(user.accessToken.split('.')[1], 'base64').toString('utf-8'));
+            }
+            return token;
+        },
+        async session({ session, token, user }) {
+            session.accessToken = token.accessToken;
+            return session;
+        }
+    }
+};
